// src/app/model/tbl-top4-offers.ts
export class tbl_top_4_offers{

    offer_id: number | undefined; 
    off_id: number | undefined; 
    restaurant_id: number | undefined; 
    title: number | undefined; 
    description: number | undefined; 
    from_date: number | undefined; 
    to_date: number | undefined; 
    booking_from: number | undefined; 
    booking_to: number | undefined; 
    capacity: number | undefined; 
    chief_guest: number | undefined; 
    price: number | undefined; 
    sgst: number | undefined; 
    cgst: number | undefined; 
    service_charge: number | undefined; 
    discount: number | undefined; 
    discount_type: number | undefined; 
    final_price: number | undefined; 
    bg_image: number | undefined; 
    create_date: number | undefined; 
    create_by: number | undefined; 
    update_date: number | undefined; 
    update_by: number | undefined; 
    status: number | undefined; 
    restaurant_name: number | undefined; 
    country_en: number | undefined; 
    state_en: string | undefined; 
    city_en: string| undefined; 

}