// src/app/model/tbl-top-8-items.ts
export class tbl_top_8_items {
        
          item_id: number | undefined;
          item: string | undefined;
          restaurant_id: number | undefined;
          category: number | undefined;
          sub_category: number | undefined;
          item_group: number | undefined;
          item_type: number | undefined;
          description: string | undefined;
          full_price: number | undefined;
          medium_price: number | undefined;
          small_price: number | undefined;
          image_item: string |undefined;
          status: number | undefined;
          unit: number | undefined;
          restaurant_name: string | undefined;
          country_en: string | undefined;
          state_en: string | undefined;
          city_en: string | undefined;
}