// src/app/model/tbl-top-12-restaurants-list.ts
export class  tbl_top_12_restaurants_list {

          id: number | undefined;
          rest_id: string | undefined;
          restaurant_id: number | undefined;
          restaurant_name: string | undefined;
          address: string | undefined;
          country_id: number | undefined;
          state_id: number | undefined;
          city_id: number | undefined;
          gst: number | undefined;
          pan: number | undefined;
          latitude: number | undefined;
          longitude: number | undefined;
          owner_name: number | undefined;
          logo: string | undefined;
          bg_image: string | undefined;
          create_date: Date | undefined;
          create_by: Date | undefined;
          update_date: Date | undefined;
          update_by: Date | undefined;
          status: number | undefined;
          country_en: string | undefined;
          state_en: string | undefined;
          city_en: string | undefined;
          discount:number | undefined;
          discount_type: number | undefined;
          discount_title: string | undefined;  
}